move job module to job package
