+++ src/osekit/public/project.py
@@ -39,7 +39,7 @@ if TYPE_CHECKING:
     from pandas import Timestamp
 
     from osekit.core.audio_file import AudioFile
-    from osekit.utils.job import JobBuilder
+    from osekit.job.job import JobBuilder
 
 
 class Project:
